+++ components/SettingsUI.tsx/SettingItem.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 interface SettingItemProps {
 	name: string;
-	description: string;
+	description?: string;
 	vertical?: boolean;
 	children: React.ReactNode;
 }
@@ -16,7 +16,9 @@ export const SettingItem: React.FC<SettingItemProps> = ({
 	<div className={`setting-item ${vertical && "setting-item-vertical"}`}>
 		<div className="setting-item-info">
 			<div className="setting-item-name">{name}</div>
-			<div className="setting-item-description">{description}</div>
+			{description && (
+				<div className="setting-item-description">{description}</div>
+			)}
 		</div>
 		<div
 			className={vertical ? "setting-item-full" : "setting-item-control"}
